organized logging scripts
