+++ XRTK-Core/Packages/com.xrtk.core/Editor/BuildPipeline/Logging.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dfc0a79f7879dd447b1ea0e2c0ad464e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
