+++ XRTK-Core/Packages/com.xrtk.core/Editor/BuildPipeline/Logging/AbstractCILogger.cs.meta
