+++ XRTK-Core/Packages/com.xrtk.core/Editor/BuildPipeline/Logging/CILoggingUtility.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7c4098916cb84858804fea3e9fab5b0c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 8ac5213854cf4dbabd140decf8df1946, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
