+++ XRTK-Core/Packages/com.xrtk.core/Editor/BuildPipeline/Logging/GitHubActionsLogger.cs
@@ -1,0 +1,15 @@
+﻿// Copyright (c) XRTK. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+namespace XRTK.Editor.BuildPipeline.Logging
+{
+    /// <summary>
+    /// https://docs.github.com/en/actions/learn-github-actions/workflow-commands-for-github-actions#about-workflow-commands
+    /// </summary>
+    public class GitHubActionsLogger : AbstractCILogger
+    {
+        public override string Error => "::error::";
+
+        public override string Warning => "::warning::";
+    }
+}
