+++ XRTK-Core/Packages/com.xrtk.core/Editor/BuildPipeline/Logging/GitHubActionsLogger.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 619084c1c97e5e04cb55a1304a5292c0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 8ac5213854cf4dbabd140decf8df1946, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
