+++ XRTK-Core/Packages/com.xrtk.core/Editor/BuildPipeline/Logging/ICILogger.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 74ad7b8b4d984c1a99d159158249e4d8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 8ac5213854cf4dbabd140decf8df1946, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
