+++ XRTK-Core/Packages/com.xrtk.core/Editor/BuildPipeline/UnityPlayerBuildTools.cs
@@ -14,6 +14,7 @@ using UnityEditor.Build;
 using UnityEditor.Build.Reporting;
 using UnityEngine;
 using XRTK.Attributes;
+using XRTK.Editor.BuildPipeline.Logging;
 using XRTK.Editor.Extensions;
 using XRTK.Editor.Utilities;
 using XRTK.Editor.Utilities.SymbolicLinks;
